Refactor - Added new module and swapped location of data fetching service
