+++ src/app/app.component.html
@@ -1,7 +1,7 @@
 <header id="store">
   <h1>Greengrocers</h1>
   <!-- your store here -->
-  <app-item-list/>
+  <app-items/>
 </header>
 <main id="cart">
   <h2>Your Cart</h2>
