+++ src/app/app.module.ts
@@ -2,12 +2,12 @@ import { NgModule } from '@angular/core';
 import { BrowserModule } from '@angular/platform-browser';
 import { HttpClientModule } from "@angular/common/http";
 import { AppComponent } from './app.component';
-import { ItemListComponent } from './item-list/item-list.component';
 import { CustomerCartModule } from './customer-cart/customer-cart.module';
+import { ItemListModule } from './item-list/item-list.module';
 
 @NgModule({
-  declarations: [AppComponent, ItemListComponent],
-  imports: [BrowserModule, HttpClientModule, CustomerCartModule],
+  declarations: [AppComponent],
+  imports: [BrowserModule, HttpClientModule, CustomerCartModule, ItemListModule],
   providers: [],
   bootstrap: [AppComponent],
 })
