+++ src/app/item-list/datafetching.service.ts
@@ -2,7 +2,7 @@ import { Injectable } from '@angular/core';
 import { HttpClient } from '@angular/common/http';
 import { Observable } from "rxjs";
 import { environment } from 'src/environments/environment.development';
-import { Item } from './models/item';
+import { Item } from '../models/item';
 
 @Injectable({
   providedIn: 'root'
